+++ src/lib/components/cosmos-evm-txs/types.ts
@@ -1,0 +1,4 @@
+export enum CosmosEvmTxsTab {
+  Cosmos = "cosmos",
+  Evm = "evm",
+}