+++ src/lib/pages/account-details/components/tables/txs/sequencer.tsx
@@ -2,6 +2,8 @@ import type { BechAddr20 } from "lib/types";
 
 import { Box, Flex } from "@chakra-ui/react";
 import { useEvmConfig, useMobile } from "lib/app-provider";
+import { CosmosEvmTxs } from "lib/components/cosmos-evm-txs";
+import { CosmosEvmTxsTab } from "lib/components/cosmos-evm-txs/types";
 import { LoadNext } from "lib/components/LoadNext";
 import { EmptyState, ErrorFetching } from "lib/components/state";
 import {
@@ -11,12 +13,14 @@ import {
   ViewMore,
 } from "lib/components/table";
 import { useTxsByAddressSequencer } from "lib/services/tx";
+import { useState } from "react";
 
 import type { TxsTableProps } from "./types";
 
 export const TxsTableSequencer = ({ address, onViewMore }: TxsTableProps) => {
   const isMobile = useMobile();
   const evm = useEvmConfig({ shouldRedirect: false });
+  const [tab, setTab] = useState(CosmosEvmTxsTab.Cosmos);
 
   const {
     data,
@@ -31,17 +35,29 @@ export const TxsTableSequencer = ({ address, onViewMore }: TxsTableProps) => {
     onViewMore ? 5 : 10
   );
 
-  const title = evm.enabled ? "Cosmos transactions" : "Transactions";
+  const title = "Transactions";
   const isMobileOverview = isMobile && !!onViewMore;
 
-  return (
-    <Box mt={[4, 8]}>
-      {isMobileOverview ? (
+  if (isMobileOverview)
+    return (
+      <Box mt={[4, 8]}>
         <MobileTitle
           count={undefined}
           showCount={false}
           title={title}
           onViewMore={onViewMore}
+        />
+      </Box>
+    );
+
+  return (
+    <Box mt={[4, 8]}>
+      {evm.enabled ? (
+        <CosmosEvmTxs
+          address={address as BechAddr20}
+          setTab={setTab}
+          tab={tab}
+          type="account"
         />
       ) : (
         <Flex direction="column">